networksecurity/components/data_validation.py
from networksecurity.exception.exception import NetworkSecurityException
from networksecurity.logging.logger import logging

## configuration of the Data Ingestion Config
from networksecurity.entity.config_entity import DataValidationConfig
from networksecurity.entity.artifact_entity import DataValidationArtifact
from networksecurity.entity.artifact_entity import DataIngestionArtifact
from networksecurity.constant.training_pipeline import SCHEMA_FILE_PATH
from networksecurity.utils.main_utils.utils import read_yaml_file, write_yaml_file
import os
import sys
import numpy as np
import pandas as pd
from typing import List

from scipy.stats import ks_2samp # Used to perform the Kolmogorov-Smirnov test , check data drift


MONGO_DB_URL=os.getenv("MONGO_DB_URL")

class DataValidation:
    def __init__(self,data_ingestion_artifact:DataIngestionArtifact, data_validation_config:DataValidationConfig):
        try:
            self.data_validation_config=data_validation_config
            self.data_ingestion_artifact = data_ingestion_artifact
            self._schema_config = read_yaml_file(SCHEMA_FILE_PATH) 
        except Exception as e:
            raise NetworkSecurityException(e,sys)
        
    @staticmethod
    def read_data(file_path:str)->pd.DataFrame:
        try:
            return pd.read_csv(file_path)
        except Exception as e:
            raise NetworkSecurityException(e,sys)
        
    def validate_number_of_columns(self,df:pd.DataFrame)->bool:
        try:
            number_of_columns = len(self._schema_config['columns'])
            logging.info(f"Required number of columns: {number_of_columns}")
            logging.info(f"Dataframe has columns: {len(df.columns)}")
            if len(df.columns) == number_of_columns:
                return True
            return False
        except Exception as e:
            raise NetworkSecurityException(e,sys)
    
    def detect_dataset_drift(self,base_df,current_df,threshold=0.05)->bool:
        try:
            status = True
            report = {}
            for column in base_df.columns:
                d1 = base_df[column]
                d2 = current_df[column]

                is_same_dist = ks_2samp(d1,d2)
                p_value = is_same_dist.pvalue
                if p_value < threshold:
                    status = False
                    is_found = True
                else:
                    is_found = False
                report[column] = {
                    "p_value": p_value,
                    "drift_detected": is_found
                }
            drift_report_file_path = self.data_validation_config.drift_report_file_path

            # create directory if not exists
            dir_path = os.path.dirname(drift_report_file_path)
            os.makedirs(dir_path, exist_ok=True)
            write_yaml_file(file_path=drift_report_file_path, content=report)

            return status

        except Exception as e:
            raise NetworkSecurityException(e,sys)
    
    def initiate_data_validation(self)->DataValidationArtifact: # return type is DataValidationArtifact as ,
                                                                # after data validation we are going to create 
                                                                # DataValidationArtifact
        try:
            logging.info("Starting data validation")
            
            train_file_path = self.data_ingestion_artifact.trained_file_path
            test_file_path = self.data_ingestion_artifact.test_file_path
            
            # reading the train and test data
            train_df = DataValidation.read_data(train_file_path)
            test_df = DataValidation.read_data(test_file_path)

            # validate number of columns
            logging.info("Validating number of columns in train data")
            status = self.validate_number_of_columns(train_df)
            if(not status):
                error_message = f"Train data does not have required number of columns"
            
            status = self.validate_number_of_columns(test_df)
            if(not status):
                error_message = f"Test data does not have required number of columns"
            
            #checking numerical columns
            numerical_columns = self._schema_config['numerical_columns']
            numerical_columns_in_train = train_df.select_dtypes(include=['int64','float64']).columns.tolist()
            numerical_columns_in_test = test_df.select_dtypes(include=['int64','float64']).columns.tolist()
            
            if(len(numerical_columns) != len(numerical_columns_in_train)):
                error_message = f"Train data does not have all numerical columns"
            if(len(numerical_columns) != len(numerical_columns_in_test)):
                error_message = f"Test data does not have all numerical columns"

            # Checking data drift
            status = self.detect_dataset_drift(base_df=train_df,current_df=test_df)
            dir_path = os.path.dirname(self.data_validation_config.valid_train_file_path)
            os.makedirs(dir_path, exist_ok=True)
            train_df.to_csv(self.data_validation_config.valid_train_file_path,index=False,header=True)
            test_df.to_csv(self.data_validation_config.valid_test_file_path,index=False,header=True)

            data_validation_artifact = DataValidationArtifact(
                validation_status = status,
                valid_train_file_path = self.data_validation_config.valid_train_file_path,
                valid_test_file_path = self.data_validation_config.valid_test_file_path,
                invalid_train_file_path= None,
                invalid_test_file_path = None,
                drift_report_file_path = self.data_validation_config.drift_report_file_path
            )

            logging.info(f"Data validation artifact: {data_validation_artifact}")
            return data_validation_artifact
        except Exception as e:
            raise NetworkSecurityException(e,sys)
    
